HDDS-13160. Remove SnapshotDirectoryCleaningService and refactor AbstractDeletingService
